% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix2mgsem_s2.R
\name{MixMix_Step2}
\alias{MixMix_Step2}
\title{Run Step 2: structural relations clustering}
\usage{
MixMix_Step2(
  s1out,
  step2model,
  nclus,
  nfree,
  seed = NULL,
  userStart = NULL,
  max_it = 10000L,
  nstarts = 50L,
  printing = FALSE,
  partition = "hard",
  Endo2Cov = TRUE,
  allG = TRUE
)
}
\arguments{
\item{s1out}{Output from \code{\link[=MixMix_Step1]{MixMix_Step1()}}.}

\item{step2model}{A lavaan model syntax string specifying the structural relations among latent variables.}

\item{nclus}{Integer scalar indicating the number of SR-clusters.}

\item{nfree}{Integer scalar indicating the number of free non-marker loadings in the Step 1 measurement model. This is used to calculate the number of parameters involved.}

\item{seed}{Optional integer seed before generating the random starts.}

\item{userStart}{Optional user-defined cluster membership matrix (dimensions: number of groups × number of clusters) with binary values
(1 or 0; 1 indicates cluster membership). This can be used to initialize the EM algorithm when the user has prior information
about cluster memberships. There must be only one \code{1} for each row. Skips random starts.}

\item{max_it}{Maximum number of EM iterations per random start.}

\item{nstarts}{Number of random starts (to avoid local maxima).}

\item{printing}{Logical indicating whether to print iteration progress.}

\item{partition}{Character string specifying the initial partition type. Currently \code{"hard"} and \code{"soft"} are supported.}

\item{Endo2Cov}{Logical indicating whether to allow covariances between endogenous 2 variables.}

\item{allG}{Logical indicating whether the endogenous covariances are group-specific (\code{TRUE}) or cluster-specific (\code{FALSE}).}
}
\value{
A named list containing:
\describe{
\item{posteriors}{A matrix of posterior SR-cluster membership probabilities.}
\item{final_fit}{The final lavaan fit object (containing all group-cluster combinations).}
\item{param}{A list containing the estimated group- and cluster-specific residual covariance matrices and cluster-specific regression matrices.}
\item{logLik}{A list of log-likelihood values from the final solution and random starts.}
\item{model_sel}{Model selection indices including BIC, AIC, and AIC3.}
\item{NrPar}{The number of parameters used for model selection.}
\item{step2_time}{Step 2 computation time in minutes.}
}
}
\description{
Fit the second step of the double-mixture multigroup SEM (2MixMG-SEM) method.
Given the group-specific factor covariance matrices obtained from \code{\link[=MixMix_Step1]{MixMix_Step1()}},
groups are clustered based on their structural relations among latent variables.
An EM algorithm is used to optimize this log-likelihood function. In the E-step, the
algorithm estimates the expected values of the cluster memberships of the groups given
the current parameter estimates. In the M-step, the algorithm maximizes the unknown
parameters given the expected cluster memberships from the E-step by calling lavaan.
}
\details{
The implementation is adapted from Andres Felipe Perez Alonso's Mixture multigroup SEM code \code{mmgsem} and extended for the step-wise 2MixMG-SEM method.
Model-selection indices are calculated using both the factor-level likelihood from Step 2 and an observed data likelihood that combines the MM-cluster and SR-cluster solutions.
}
\examples{
\dontrun{
step2_model <- '
  F4 ~ F1 + F3
  F3 ~ F1 + F2'

step2_out <- MixMix_Step2(
  s1out = step1_out,
  step2model = step2_model,
  nclus = 2,
  nfree = 4,
  seed = 100
)
}
}
