% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix2mgsem_s2_modsel.R
\name{print.mix2mgsem_step2_selection}
\alias{print.mix2mgsem_step2_selection}
\title{Print a Step 2 model selection object}
\usage{
\method{print}{mix2mgsem_step2_selection}(x, ...)
}
\arguments{
\item{x}{An object returned by \code{\link[=MixMix_Step2_select]{MixMix_Step2_select()}}.}

\item{...}{Additional arguments passed to \code{print.data.frame()}.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a Step 2 model selection object
}
