% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix2mgsem_s2_modsel.R
\name{MixMix_Step2_select}
\alias{MixMix_Step2_select}
\title{Select the number of SR-clusters for Step 2}
\usage{
MixMix_Step2_select(s1out, step2model, nclus, ...)
}
\arguments{
\item{s1out}{Output from \code{\link[=MixMix_Step1]{MixMix_Step1()}}.}

\item{step2model}{A lavaan model syntax string specifying the structural relations among latent variables.}

\item{nclus}{Integer vector containing at least two candidate numbers of SR-clusters.}

\item{...}{Additional arguments passed to \code{\link[=MixMix_Step2]{MixMix_Step2()}}.}
}
\value{
An object of class \code{mix2mgsem_step2_selection} containing:
\describe{
\item{fits}{A list of \code{\link[=MixMix_Step2]{MixMix_Step2()}} fits. Failed fits are stored as \code{NULL}.}
\item{overview}{Data frame containing the fit and information criteria for the fitted models.}
\item{selected}{Numbers of SR-clusters selected by each information criterion.}
}
}
\description{
Fit \code{\link[=MixMix_Step2]{MixMix_Step2()}} for several numbers of SR-clusters and compare the model fit.
}
\examples{
\dontrun{
s2_selection <- MixMix_Step2_select(
  s1out = s1out,
  step2model = step2_model,
  nclus = c(1:5),
  nfree = 4,
  seed = 100
)

s2_selection$overview

plot(
  s2_selection,
  criteria = c("BIC_G", "BIC_N", "AIC", "AIC3")
)
}
}
